+++ pystruct3d/annotation/__init__.py
@@ -1,0 +1,3 @@
+from pystruct3d.annotation.utils import transfer_labels as transfer_labels
+
+__all__ = ["transfer_labels"]