+++ pystruct3d/bbox/__init__.py
@@ -1,0 +1,5 @@
+from pystruct3d.bbox.bbox import BBox as BBox
+from pystruct3d.bbox.utils import bbox_array2list as bbox_array2list
+from pystruct3d.bbox.utils import bbox_list2array as bbox_list2array
+
+__all__ = ["BBox", "bbox_array2list", "bbox_list2array"]