+++ pystruct3d/preprocessing/__init__.py
@@ -4,3 +4,16 @@ from pystruct3d.preprocessing.labels import filter_ids as filter_ids
 from pystruct3d.preprocessing.labels import labels_to_color as labels_to_color
 from pystruct3d.preprocessing.pca import rotate_by_pca as rotate_by_pca
 from pystruct3d.preprocessing.pca import simple_pca as simple_pca
+from pystruct3d.preprocessing.voxel import density_filter as density_filter
+from pystruct3d.preprocessing.voxel import downsample as downsample
+
+__all__ = [
+    "align_to_axes",
+    "crop_roi",
+    "density_filter",
+    "downsample",
+    "filter_ids",
+    "labels_to_color",
+    "rotate_by_pca",
+    "simple_pca",
+]